+++ packages/ui-causal-graph-editor/src/shared/editor-overlay/panel-content/edge/sections/constraint-editor.tsx
@@ -14,6 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 import { ButtonBar, Item, Tooltip } from '@darajs/ui-components';
 import { ArrowRightLong, ArrowsHorizontal, Ban } from '@darajs/ui-icons';
 
@@ -47,7 +48,7 @@ const constraintItems: IconItem[] = [
                 </span>
             </Tooltip>
         ),
-        value: EdgeConstraintType.DIRECTED,
+        value: EdgeConstraintType.HARD_DIRECTED,
     },
     {
         label: (