+++ packages/ui-causal-graph-editor/src/types.tsx
@@ -200,14 +200,13 @@ export type SimulationGraph = AbstractGraph<SimulationNode, SimulationEdge, Simu
 /**
  * Type of an edge constraint that can be encoded
  *
- * Int values from cl_interfaces.EdgeConstraint
+ * str values from cai_causal_graph.EdgeConstraint
  */
 export enum EdgeConstraintType {
-    DIRECTED = 2,
-    /** Backward directed - unsupported */
-    BACKWARD_DIRECTED = 3,
-    FORBIDDEN = 4,
-    UNDIRECTED = 1,
+    FORBIDDEN = 'forbidden',
+    HARD_DIRECTED = 'hard_directed',
+    SOFT_DIRECTED = 'soft_directed',
+    UNDIRECTED = 'hard_undirected',
 }
 
 /**